Move MVStoreBase to logic.util.io to decouple persistence from AI package
