+++ jablib/src/main/java/org/jabref/logic/ai/summarization/repositories/MVStoreSummariesRepository.java
@@ -4,9 +4,9 @@ import java.nio.file.Path;
 import java.util.Map;
 import java.util.Optional;
 
-import org.jabref.logic.ai.util.MVStoreBase;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.util.NotificationService;
+import org.jabref.logic.util.io.MVStoreBase;
 import org.jabref.model.ai.summarization.AiSummary;
 import org.jabref.model.ai.summarization.AiSummaryIdentifier;
